requirements.txt
certifi==2021.10.8
cycler==0.11.0
dlib==19.23.1
fonttools==4.31.2
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==3.5.1
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
plotly==5.7.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sklearn==0.0
tenacity==8.0.1
threadpoolctl==3.1.0